requirements.txt
pydantic>=2.0.0
pydantic-graph>=0.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-fasthtml>=0.1.0
httpx>=0.26.0
python-multipart>=0.0.6
itsdangerous>=2.1.2
python-dotenv>=1.0.0
chromadb>=0.4.0
openai>=1.0.0
